src/fairness.py
```python
# Importing necessary libraries
from fairlearn.metrics import MetricFrame
from fairlearn.reductions import ExponentiatedGradient, DemographicParity
from aif360.algorithms.preprocessing import Reweighing
from aif360.datasets import BinaryLabelDataset
from themis_ml.metrics import mean_difference
from themis_ml.postprocessing import Relabeller
from auditai.fairness import check_algorithmic_fairness

# Define a function to check fairness of a model
def check_fairness(y_true, y_pred, sensitive_features):
    """
    This function checks the fairness of a model's predictions using Fairlearn's MetricFrame.
    """
    mf = MetricFrame(metrics={'accuracy': accuracy_score,
                              'precision': precision_score,
                              'recall': recall_score,
                              'false positive rate': false_positive_rate,
                              'false negative rate': false_negative_rate,
                              'selection rate': selection_rate,
                              'demographic parity difference': demographic_parity_difference,
                              'demographic parity ratio': demographic_parity_ratio,
                              'equalized odds difference': equalized_odds_difference,
                              'equalized odds ratio': equalized_odds_ratio},
                     y_true=y_true,
                     y_pred=y_pred,
                     sensitive_features=sensitive_features)
    return mf.difference()

# Define a function to mitigate bias using Reweighing from AI Fairness 360
def mitigate_bias(dataset, protected_attribute_names):
    """
    This function mitigates bias in a dataset using Reweighing from AI Fairness 360.
    """
    rw = Reweighing(unprivileged_groups=[{'race': 0}],
                    privileged_groups=[{'race': 1}])
    dataset_transf = rw.fit_transform(dataset)
    return dataset_transf

# Define a function to check algorithmic fairness using Audit-AI
def check_algorithmic_fairness(y_true, y_pred):
    """
    This function checks the algorithmic fairness of a model's predictions using Audit-AI.
    """
    results = check_algorithmic_fairness(y_true, y_pred)
    return results
```
