tests/test_fairness.py
```python
# Importing necessary libraries
import unittest
import numpy as np
from src.fairness import check_fairness, mitigate_bias, check_algorithmic_fairness

class TestFairness(unittest.TestCase):

    def setUp(self):
        # Create some dummy data for testing
        self.y_true = np.array([1, 0, 1, 1, 0, 1, 0, 0, 1, 1])
        self.y_pred = np.array([1, 1, 0, 1, 0, 1, 0, 0, 1, 0])
        self.sensitive_features = np.array(['male', 'female', 'male', 'female', 'male', 'female', 'male', 'female', 'male', 'female'])

    def test_check_fairness(self):
        # Test the check_fairness function
        result = check_fairness(self.y_true, self.y_pred, self.sensitive_features)
        self.assertIsInstance(result, dict)

    def test_mitigate_bias(self):
        # Test the mitigate_bias function
        dataset = BinaryLabelDataset(df=pd.DataFrame(self.y_true, columns=['label']),
                                     protected_attribute_names=['gender'],
                                     favorable_label=1.0,
                                     unfavorable_label=0.0,
                                     unprivileged_protected_attributes=['male'],
                                     privileged_protected_attributes=['female'])
        result = mitigate_bias(dataset, ['gender'])
        self.assertIsInstance(result, BinaryLabelDataset)

    def test_check_algorithmic_fairness(self):
        # Test the check_algorithmic_fairness function
        result = check_algorithmic_fairness(self.y_true, self.y_pred)
        self.assertIsInstance(result, dict)

if __name__ == '__main__':
    unittest.main()
```
